Fixed adding answers and flipped sA -> essay relationship. Again.
