+++ src/completedResponse/SurveyResponse.java
@@ -1,5 +1,5 @@
 package completedResponse;
 
 public class SurveyResponse {
-
+	public SurveyResponse(){}
 }
