+++ src/completedResponse/TestGrade.java
@@ -1,5 +1,5 @@
 package completedResponse;
 
 public class TestGrade {
-
+	public TestGrade(){};
 }
